pyproject.toml
[project]
name = "pyramid-kvs"
version = "1.0.0"
description = "Session and cache for Pyramid"
authors = [{ name = "Gandi", email = "tech@gandi.net" }]
readme = "README.rst"
requires-python = ">=3.7"
license = { text = "GNU" }
dependencies = [
    "pyramid >=1.10.0,<3",
    "redis >=4.0,<6.0",
    "python3-memcached >=1.50",
    "importlib_metadata; python_version <'3.8'",
]
classifiers = [
    "Programming Language :: Python",
    "Framework :: Pyramid",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: WSGI",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",

]

[project.urls]
Homepage = "https://github.com/Gandi/pyramid_kvs"
Documentation = "https://github.com/Gandi/pyramid_kvs/blob/main/README.rst"
Repository = "https://github.com/Gandi/pyramid_kvs.git"
Issues = "https://github.com/Gandi/pyramid_kvs/issues"
Changelog = "https://github.com/Gandi/pyramid_kvs/blob/main/CHANGES.rst"


[dependency-groups]
dev = [
    "black ==23.3.0",
    "flake8 >=5,<6; python_version <'3.9'",
    "flake8 >=7; python_version >='3.10'",
    "flake8-formatter-junit-xml >=0.0.6,<1",
    "isort >=5.10.1,<6",
    "pytest >=7.2.2,<8",
    "pytest-cov >=4.0.0,<5",
]

[tool.pdm.build]
includes = ["src", "CHANGELOG.src"]
excludes = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py37"

[tool.ruff.lint]
select = [
    "B",   # bug bear security warning
    "I",   # isort import order
    "F",   # pyflakes
    "UP",  # alter when better syntax is available
    "RUF", #  the ruff devleoper's own rules
]
ignore = ["UP006", "UP035", "RUF012"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
